MutantOutbreak(Beta)
VERSION 1.3
- Se arreglan errores e n el menu
- Se arreglan algunas animaciones
- Se añade un sistema de municion
- Se añaden "Perks" o invocaciones
- Se añaden nuevos efectos de sonido

+++ Mutant Outbreak/Proyecto/Assets/Scripts/BarraDeVidaZombie.cs
@@ -17,7 +17,7 @@ public class BarraDeVidaZombie : MonoBehaviour
         }
     }
 
-    public void ActualizarVida(int vidaActual, int vidaMaxima)
+    public void ActualizarVida(float vidaActual, float vidaMaxima)
     {
         if (sliderVida != null)
         {
